refactor into additional files, fix some tests
