+++ mlflow/genai/agent_server/types.py
@@ -1,0 +1,5 @@
+"""Type definitions for the agent server module."""
+
+from typing import Literal
+
+AgentType = Literal["agent/v1/responses", "agent/v1/chat", "agent/v2/chat"]
